Add files via upload
Our task was to find if it is possible to cut the cake in the below
mentioned ways for a given value of N.
Given an integer N and a cake which can be cut into pieces, each
cut should be a straight line going from the center of the cake to
its border. Also, the angle between any two cuts must be a
positive integer. Two pieces are equal if their appropriate angles
are equal.
The given cake can be cut in following three ways:
• Cut the cake into N equal pieces.
• Cut the cake into N pieces of any size.
• Cut the cake into N pieces such that no two of them are equal